+++ MLX42/src/mlx_mouse.c
@@ -6,7 +6,7 @@
 /*   By: W2Wizard <w2.wizzard@gmail.com>              +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/01/01 23:20:13 by W2Wizard      #+#    #+#                 */
-/*   Updated: 2022/04/13 00:26:43 by w2wizard      ########   odam.nl         */
+/*   Updated: 2022/05/03 15:36:17 by mgraaf        ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 //= Private =//
 
-static void mlx_scroll_cb(GLFWwindow* window, double xoffset, double yoffset)
+static void mlx_scroll_cb(GLFWwindow* window, float xoffset, float yoffset)
 {
 	const mlx_t* mlx = glfwGetWindowUserPointer(window);
 	const mlx_scroll_t scroll_hook = ((mlx_ctx_t*)mlx->context)->scroll_hook;
@@ -30,7 +30,7 @@ static void mlx_mouse_cb(GLFWwindow* window, int32_t button, int32_t action, int
 	mouse_hook.func(button, action, mods, mouse_hook.param);
 }
 
-static void mlx_cursor_cb(GLFWwindow* window, double xpos, double ypos)
+static void mlx_cursor_cb(GLFWwindow* window, float xpos, float ypos)
 {
 	const mlx_t* mlx = glfwGetWindowUserPointer(window);
 	const mlx_cursor_t cursor_hook = ((mlx_ctx_t*)mlx->context)->cursor_hook;
@@ -84,7 +84,7 @@ void mlx_set_mouse_pos(mlx_t* mlx, int32_t x, int32_t y)
 {
 	MLX_ASSERT(!mlx);
 
-	glfwSetCursorPos(mlx->window, (double)x, (double)y);
+	glfwSetCursorPos(mlx->window, (float)x, (float)y);
 }
 
 void mlx_get_mouse_pos(mlx_t* mlx, int32_t* x_out, int32_t* y_out)
@@ -93,7 +93,7 @@ void mlx_get_mouse_pos(mlx_t* mlx, int32_t* x_out, int32_t* y_out)
 	MLX_ASSERT(!x_out);
 	MLX_ASSERT(!y_out);
 
-	double	x, y;
+	float	x, y;
 	glfwGetCursorPos(mlx->window, &x, &y);
 	*x_out = (int32_t)x;
 	*y_out = (int32_t)y;
